requirements.txt
# Polymarket Bot Requirements
# تثبيت: pip install -r requirements.txt

# HTTP & API
httpx>=0.24.0
requests>=2.28.0

# Environment
python-dotenv>=1.0.0

# Web3 & Blockchain
web3>=6.0.0
eth-account>=0.10.0

# Polymarket
py-clob-client>=0.40.0
py-order-utils>=0.1.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Async
asyncio>=3.4.3
aiohttp>=3.8.0

# Logging
colorlog>=6.7.0

# Date & Time
python-dateutil>=2.8.0

# Optional: Telegram
# python-telegram-bot>=20.0

# Optional: Database
# sqlalchemy>=2.0.0
# sqlite3
